feat: Add --unset flag to unset the default owner
This commit adds a new `--unset` flag to the `gh owner` command, allowing users to unset the default owner. This is useful when the default owner needs to be removed or changed. The `--unset` flag removes the default owner setting, allowing the user to select a new default owner in future `gh owner` commands.