feat: Refactor listOrgs function to improve performance
This commit refactors the `listOrgs` function in the `http.go` file to improve performance. The function is renamed to `listAllOrgs` to better reflect its purpose. Additionally, a cached HTTP client is used to reduce the number of API requests made. This optimization will result in faster retrieval of organization data when calling the `gh owner` command.