feat: Update cachedClient expiration time to one week
This commit updates the `cachedClient` initialization in the `http.go` file to set the expiration time to one week (`time.Hour*24*7`). By extending the expiration time, we can reduce the frequency of API requests made by the `listAllOrgs` function, resulting in improved performance and faster retrieval of organization data when using the `gh owner` command.