Refactor unarchive command to use DefaultOwner field
This commit refactors the unarchive command in the unarchive.go file to use the DefaultOwner field from the UnarchiveOptions struct. The DefaultOwner field is a function that retrieves the default owner from the configuration. This change enables the use of the default owner when unarchiving repositories, improving flexibility and convenience.