Refactor archive command to use DefaultOwner field
This commit refactors the archive command in the archive.go file to use the DefaultOwner field from the ArchiveOptions struct. The DefaultOwner field is a function that retrieves the default owner from the configuration. By using this field, the code ensures that the owner is set to the default value when no owner is provided as an argument. This change improves flexibility and convenience when archiving repositories.