Refactor forkRun function to use DefaultOwner field
This commit refactors the forkRun function in the fork.go file to use the DefaultOwner field from the ForkOptions struct. The DefaultOwner field is a function that retrieves the default owner from the configuration. By using this field, the code ensures that the owner is set to the default value when no owner is provided as an argument. This change improves flexibility and convenience when forking repositories.